backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
yt-dlp==2023.11.16
pandas==2.1.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0 